// src/flatbuffers/ctf/message-type.ts
// automatically generated by the FlatBuffers compiler, do not modify

export enum MessageType{
  PlayerJoin = 1,
  PlayerLeave = 2,
  MatchStart = 3,
  MatchState = 4,
  ClientMove = 5,
  ClientFire = 6,
  ClientCancelFire = 7,
  ClientChargeFire = 8,
  PlayerMove = 9,
  PlayerChargeFire = 10,
  PlayerFire = 11,
  PlayerCancelFire = 12,
  PlayerKilled = 13,
  PlayerRespawn = 14,
  FlagTaken = 15,
  FlagDropped = 16,
  FlagReturned = 17,
  TeamScored = 18,
  MatchEnd = 19
}